requirements.txt
appdirs==1.4.3
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
black==19.10b0
CacheControl==0.12.6
certifi==2019.11.28
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.1
distro==1.4.0
Django==3.0.8
django-debug-toolbar==2.2
django-extensions==3.0.3
djangorestframework==3.11.0
drf-nested-routers==0.91
factory-boy==2.12.0
Faker==4.1.1
filelock==3.0.12
flake8==3.8.3
flake8-bugbear==20.1.4
flake8-django==1.1.1
html5lib==1.0.1
identify==1.4.25
idna==2.8
importmagic3==0.2.0
ipaddr==2.2.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
lockfile==0.12.2
log-colorizer==1.8.6
mccabe==0.6.1
more-itertools==8.4.0
msgpack==0.6.2
nodeenv==1.4.0
packaging==20.3
parso==0.7.0
pathspec==0.8.0
pep517==0.8.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
pre-commit==2.6.0
progress==1.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.6
pytest==5.4.3
pytest-django==3.9.0
python-dateutil==2.8.1
python-magic==0.4.18
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.22.0
retrying==1.3.3
six==1.14.0
spark-parser==1.8.9
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
uncompyle6==3.7.3
urllib3==1.25.8
virtualenv==20.0.28
wcwidth==0.2.5
wdb==3.3.0
webencodings==0.5.1
xdis==5.0.3
